put-reservas

+++ slim/src/reservasController.php
@@ -4,20 +4,21 @@ require_once __DIR__ . "/../utils/utils.php";
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 
-function getReservas (Request $request, Response $response) {
-    try{
+function getReservas(Request $request, Response $response)
+{
+    try {
         $pdo = getConnection();
         $sql = "SELECT * FROM reservas";
         $consulta = $pdo->query($sql);
         $resultados = $consulta->fetchAll(PDO::FETCH_ASSOC);
-        $payload = json_encode ([
+        $payload = json_encode([
             'status' => 'success',
             'code' => 200,
             'data' => $resultados
         ]);
         $response->getBody()->write($payload);
-        return $response->withStatus(200); 
-    } catch (\Exception $e){
+        return $response->withStatus(200);
+    } catch (\Exception $e) {
         $payload = json_encode([
             'code' => '500',
             'error' => $e->getMessage()
@@ -27,33 +28,34 @@ function getReservas (Request $request, Response $response) {
     }
 };
 
-function postReservas (Request $request, Response $response) {
+function postReservas(Request $request, Response $response)
+{
     $data = $request->getParsedBody();
     $requiredFields = ['propiedad_id', 'inquilino_id', 'fecha_inicio_disponibilidad', 'cantidad_noches'];
-    $responseVal = validationFields($data,$requiredFields,$response);
-    if (!$responseVal){
+    $responseVal = validationFields($data, $requiredFields, $response);
+    if (!$responseVal) {
         return $response->withStatus(400);
     } else {
-        try{
+        try {
             $pdo = getConnection();
 
             $inquilino_id = $data['inquilino_id'];
-            $sql = "SELECT activo FROM inquilinos WHERE id = '". $inquilino_id ."'";
+            $sql = "SELECT activo FROM inquilinos WHERE id = '" . $inquilino_id . "'";
             $consulta = $pdo->query($sql);
             $resultado = $consulta->fetchAll(PDO::FETCH_ASSOC);
             $active = $resultado[0]['activo'];
 
             $propiedad_id = $data['propiedad_id'];
-            $sql = "SELECT disponible FROM propiedades WHERE id = '". $propiedad_id ."'";
+            $sql = "SELECT disponible FROM propiedades WHERE id = '" . $propiedad_id . "'";
             $consulta = $pdo->query($sql);
             $resultado = $consulta->fetchAll(PDO::FETCH_ASSOC);
             $disponible = $resultado[0]['disponible'];
-            if ($active == 1 && $disponible == 1 ) {
+            if ($active == 1 && $disponible == 1) {
                 $cantidad_noches = $data['cantidad_noches'];
-                $sql= "SELECT valor_noche FROM propiedades WHERE id = '". $propiedad_id ."'";
+                $sql = "SELECT valor_noche FROM propiedades WHERE id = '" . $propiedad_id . "'";
                 $consulta = $pdo->query($sql);
                 $resultado = $consulta->fetchAll(PDO::FETCH_ASSOC);
-                $valor_total = $resultado[0]['valor_noche']* $cantidad_noches;
+                $valor_total = $resultado[0]['valor_noche'] * $cantidad_noches;
                 $fecha_desde = $data['fecha_desde'];
                 $sql = "INSERT INTO reservas (id,propiedad_id,inquilino_id,fecha_desde,cantidad_noches,valor_total) VALUES (:id,:propiedad_id,:inquilino_id,:fecha_desde,:cantidad_noches,:valor_total)";
                 $consulta = $pdo->prepare($sql);
@@ -78,24 +80,40 @@ function postReservas (Request $request, Response $response) {
                 $response->getBody()->write($payload);
                 return $response;
             }
-            } catch (\Exception $e){
-                $payload = json_encode([
-                    'code' => '500',
-                    'error' => $e->getMessage()
-                ]);
-                $response->getBody()->write($payload);
-                return $response->withStatus(500);
-                
-            }
+        } catch (\Exception $e) {
+            $payload = json_encode([
+                'code' => '500',
+                'error' => $e->getMessage()
+            ]);
+            $response->getBody()->write($payload);
+            return $response->withStatus(500);
+        }
     }
-    } 
+}
 
-function deleteReservas (Request $request, Response $response, $args){
+function deleteReservas(Request $request, Response $response, $args)
+{
     $id = $args['id'];
-    // validar que se encuentre en rango
-    try{
+
+    try {
         $pdo = getConnection();
-        $sql = "DELETE FROM reservas WHERE id = '". $id ."'";
+        $sql = "SELECT fecha_desde FROM reservas WHERE id = '" . $id . "'";
+        $consulta = $pdo->query($sql);
+        $resultado = $consulta->fetchAll(PDO::FETCH_ASSOC);
+        $fecha_desde = $resultado[0]['fecha_desde'];
+        $fecha_actual = date("Y-m-d");
+        $fecha1 = new DateTime($fecha_desde);
+        $fecha2 = new DateTime($fecha_actual);
+        // var_dump($resultado[0]['fecha_desde'],$fecha_actual);die;
+        if ($fecha_desde < $fecha_actual) {
+            $payload = json_encode([
+                'code' => 201,
+                'message' => 'La reserva no se puede eliminar una vez iniciada la estadía'
+            ]);
+            $response->getBody()->write($payload);
+            return $response->withStatus(201);
+        }
+        $sql = "DELETE FROM reservas WHERE id = '" . $id . "'";
         $consulta = $pdo->query($sql);
         $payload = json_encode([
             'code' => 201,
@@ -113,44 +131,110 @@ function deleteReservas (Request $request, Response $response, $args){
     }
 }
 
-function putReservas (Request $request, Response $response, $args){
-    // validar que se encuentre en rango, si comenzo a correr los dias
+function putReservas(Request $request, Response $response, $args)
+{
     $data = $request->getParsedBody();
     $id = $args['id'];
-    // $propiedad_id = isset($data['propiedad_id']) ? $data['propiedad_id'] : null;
-    // $inquilino_id = isset($data['inquilino_id']) ? $data['inquilino_id'] : null;
-    // $fecha_desde = isset($data['fecha_desde']) ? $data['fecha_desde'] : null;
-    // $cantidad_noches = isset($data['cantidad_noches']) ? $data['cantidad_noches'] : null;
-    $valor_total = isset($data['valor_total']) ? $data['valor_total'] : null;
-
-    // if (!isset($data['inquilino_id']) || !isset($data['fecha_desde']) || !isset($data['cantidad_noches']) || !isset($data['valor_total'])) {
-    //     $payload = json_encode(['error' => 'Faltan completar campos obligatorios', 'code' => '400']);
-    //     $response->getBody()->write($payload);
-    //     return $response->withStatus(400);
-    // }
-    //else {
-        try {
-            $pdo = getConnection();
-            $sql = "UPDATE reservas SET valor_total = (:valor_total) WHERE id = (:id)";
-            $consulta = $pdo->prepare($sql);
-            $consulta->bindValue(':valor_total', $valor_total, PDO::PARAM_STR);
-            $consulta->bindValue(':id', $id, PDO::PARAM_INT);
-            $consulta->execute();
+    try {
+        $pdo = getConnection();
+        $sql = "SELECT fecha_desde FROM reservas WHERE id = '" . $id . "'";
+        $consulta = $pdo->query($sql);
+        $resultado = $consulta->fetchAll(PDO::FETCH_ASSOC);
+        $fecha_desde = $resultado[0]['fecha_desde'];
+        $fecha_actual = date("Y-m-d");
+        $fecha1 = new DateTime($fecha_desde);
+        $fecha2 = new DateTime($fecha_actual);
+        if ($fecha1 < $fecha2) {
             $payload = json_encode([
                 'code' => 201,
-                'message' => 'Localidad editada con éxito'
+                'message' => 'La reserva no se puede modificar una vez iniciada la estadía'
             ]);
             $response->getBody()->write($payload);
             return $response->withStatus(201);
-        } catch(\Exception $e) {
-            $payload = json_encode([
-                'code' => '500',
-                'error' => $e->getMessage()
-            ]);
-            $response->getBody()->write($payload);
-            return $response->withStatus(500);
         }
-   // }
+        $data = $request->getParsedBody();
+        $propiedad_id = isset($data['propiedad_id']) ? $data['propiedad_id'] : null;
+        $inquilino_id = isset($data['inquilino_id']) ? $data['inquilino_id'] : null;
+        $fecha_inicio_disponibilidad = isset($data['fecha_inicio_disponibilidad']) ? $data['fecha_inicio_disponibilidad'] : null;
+        $cantidad_noches = isset($data['cantidad_noches']) ? $data['cantidad_noches'] : null;
+        $sql = "UPDATE reservas SET";
+        $params = [];
+        if (!empty($propiedad_id)) {
+            $sql .= " propiedad_id = :propiedad_id,";
+            $params[':propiedad_id'] = $propiedad_id;
+            $sql_2 = "SELECT valor_noche FROM propiedades WHERE id = '" . $propiedad_id . "'";
+            $consulta = $pdo->query($sql_2);
+            $resultado = $consulta->fetchAll(PDO::FETCH_ASSOC);
+            $valor_noche = $resultado[0]['valor_noche'];
+            $sql_3 = "SELECT cantidad_noches FROM reservas WHERE id = '" . $id . "'";
+            $consulta = $pdo->query($sql_3);
+            $resultado = $consulta->fetchAll(PDO::FETCH_ASSOC);
+            $cantidad_noches_aux = $resultado[0]['cantidad_noches'];
+            $valor_total = $valor_noche * $cantidad_noches_aux;
+            //var_dump($valor_total);die;
+            $sql .= " valor_total = :valor_total,";
+            $params[':valor_total'] = $valor_total;
+        }
+        
+        if (!empty($inquilino_id)) {
+            $sql .= " inquilino_id = :inquilino_id,";
+            $params[':inquilino_id'] = $inquilino_id;
+        }
+        
+        if (!empty($fecha_inicio_disponibilidad)) {
+            $sql .= " fecha_inicio_disponibilidad = :fecha_inicio_disponibilidad,";
+            $params[':fecha_inicio_disponibilidad'] = $fecha_inicio_disponibilidad;
+        }
+        // var_dump($cantidad_noches);die;
+        if (!empty($cantidad_noches)) {
+            $sql .= " cantidad_noches = :cantidad_noches,";
+            $params[':cantidad_noches'] = $cantidad_noches;
+            if (empty($propiedad_id)){
+                $sql_2 = "SELECT propiedad_id FROM reservas WHERE id = '" . $id . "'";
+                $consulta = $pdo->query($sql_2);
+                $resultado = $consulta->fetchAll(PDO::FETCH_ASSOC);
+                $propiedad_id = $resultado[0]['propiedad_id'];
+            }
+            $sql_2 = "SELECT valor_noche FROM propiedades WHERE id = '" . $propiedad_id . "'";
+            $consulta = $pdo->query($sql_2);
+            $resultado = $consulta->fetchAll(PDO::FETCH_ASSOC);
+            $valor_noche = $resultado[0]['valor_noche'];
+            $sql_3 = "SELECT cantidad_noches FROM reservas WHERE id = '" . $id . "'";
+            $consulta = $pdo->query($sql_3);
+            $resultado = $consulta->fetchAll(PDO::FETCH_ASSOC);
+            $cantidad_noches = $resultado[0]['cantidad_noches'];
+            $valor_total = $valor_noche  * $cantidad_noches;
+            $sql .= " valor_total = :valor_total,";
+            if (array_key_exists(':valor_total', $params)) {
+                // Si la clave ya existe, reemplazar su valor
+                $params[':valor_total'] = $valor_total;
+            } else {
+                // Si la clave no existe, agregarla con su valor correspondiente
+                $params[':valor_total'] = $valor_total;
+            }
+        }
+        //var_dump($valor_total);die;
+        $sql = rtrim($sql, ',');
 
+        $sql .= " WHERE id = :id";
+        $params[':id'] = $id;
+        $consulta = $pdo->prepare($sql);
+        foreach ($params as $key => $value) {
+            $consulta->bindValue($key, $value);
+        }
+        $consulta->execute();
+        $payload = json_encode([
+            'code' => 201,
+            'message' => 'Reserva editada con éxito'
+        ]);
+        $response->getBody()->write($payload);
+        return $response->withStatus(201);
+    } catch (\Exception $e) {
+        $payload = json_encode([
+            'code' => '500',
+            'error' => $e->getMessage()
+        ]);
+        $response->getBody()->write($payload);
+        return $response->withStatus(500);
+    }
 }
-
